Enhance user and task class to support logical delete etc.
